requirements.txt
cffi==1.8.3
cryptography==1.5.1
Django==1.10.1
django-bootstrap3==7.1.0
django-bootstrap3-datepicker==0.4.1
django-forms-bootstrap==3.0.1
djangorestframework==3.4.7
enum34==1.1.6
ez-setup==0.9
idna==2.1
ipaddress==1.0.17
MySQL-python==1.2.5
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pycparser==2.14
PyCronofy==1.0.0
pyOpenSSL==16.1.0
python-dateutil==2.5.3
requests==2.11.1
simple-salesforce==0.70
six==1.10.0
